requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
fastf1==3.8.1
joblib==1.5.3
scikit-learn==1.8.0
pandas==2.3.3
numpy==2.4.3
requests==2.32.5
requests-cache==1.3.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scipy==1.17.1
matplotlib==3.10.8
timple==0.1.8
RapidFuzz==3.14.3
signalrcore==1.0.2
msgpack==1.1.2
httpx==0.28.1
psycopg2-binary==2.9.11